Added some changes based on testing results

+++ workload/src/main/java/org/gym/workload/controller/TrainerWorkloadController.java
@@ -30,7 +30,7 @@ public class TrainerWorkloadController {
 
     @GetMapping
     @ResponseBody
-    public Long getDuration(@RequestParam("username") String username,
+    public int getDuration(@RequestParam("username") String username,
                             @RequestParam("year") Integer year,
                             @RequestParam("month") Integer month) {
